Wallet popup list of transactions should be per chain - coses #843
