+++ packages/tempus-client_v2/src/components/notification/NotificationContainer.tsx
@@ -1,7 +1,9 @@
 import { FC, useCallback, useContext, useEffect, useState } from 'react';
+import { Downgraded, useState as useHookState } from '@hookstate/core';
 import { UserSettingsContext } from '../../context/userSettingsContext';
 import { Notification } from '../../interfaces/Notification';
 import getNotificationService from '../../services/getNotificationService';
+import { selectedChainState } from '../../state/ChainState';
 import NotificationComponent from './NotificationComponent';
 
 import './Notification.scss';
@@ -9,6 +11,8 @@ import './Notification.scss';
 const NotificationContainer: FC = () => {
   const { setUserSettings } = useContext(UserSettingsContext);
   const [notifications, setNotifications] = useState<Notification[]>([]);
+  const selectedChain = useHookState(selectedChainState);
+  const selectedChainName = selectedChain.attach(Downgraded).get();
 
   const onNotificationDelete = useCallback(
     (id: string) => {
@@ -37,7 +41,7 @@ const NotificationContainer: FC = () => {
     const notificationStream$ = getNotificationService()
       .getNextItem()
       .subscribe(notification => {
-        if (notification) {
+        if (notification && notification.chain === selectedChainName) {
           // Create auto close timer for new notification if web page tab is visible when notification is created
           if (!document.hidden) {
             autoCloseNotification(notification.id);
@@ -48,7 +52,7 @@ const NotificationContainer: FC = () => {
       });
 
     return () => notificationStream$.unsubscribe();
-  }, [autoCloseNotification, setNotifications]);
+  }, [selectedChainName, autoCloseNotification, setNotifications]);
 
   return (
     <div className="tc__notification-container">
